cleanup(authorization): move LoadingUsersException from previous PR correctly
