+++ src/main/java/de/dhbw/karlsruhe/students/mailflow/core/application/auth/LoginUseCase.java
@@ -1,9 +1,9 @@
 package de.dhbw.karlsruhe.students.mailflow.core.application.auth;
 
 import de.dhbw.karlsruhe.students.mailflow.core.domain.auth.AuthorizationException;
+import de.dhbw.karlsruhe.students.mailflow.core.domain.auth.LoadingUsersException;
 import de.dhbw.karlsruhe.students.mailflow.core.domain.email.value_objects.Address;
 import de.dhbw.karlsruhe.students.mailflow.core.domain.user.User;
-import de.dhbw.karlsruhe.students.mailflow.external.infrastructure.authorization.LoadingUsersException;
 
 /**
  * @author seiferla
