+++ src/main/java/de/dhbw/karlsruhe/students/mailflow/core/domain/auth/LoadingUsersException.java
@@ -1,4 +1,4 @@
-package de.dhbw.karlsruhe.students.mailflow.external.infrastructure.authorization;
+package de.dhbw.karlsruhe.students.mailflow.core.domain.auth;
 
 /**
  * @author seiferla
