+++ src/main/java/de/dhbw/karlsruhe/students/mailflow/core/domain/user/UserRepository.java
@@ -1,9 +1,10 @@
 package de.dhbw.karlsruhe.students.mailflow.core.domain.user;
 
+import de.dhbw.karlsruhe.students.mailflow.core.domain.auth.LoadingUsersException;
 import de.dhbw.karlsruhe.students.mailflow.core.domain.email.value_objects.Address;
 import de.dhbw.karlsruhe.students.mailflow.core.domain.user.exceptions.SaveUserException;
-import de.dhbw.karlsruhe.students.mailflow.external.infrastructure.authorization.LoadingUsersException;
 import java.util.Optional;
+
 /**
  * @author seiferla
  */
