+++ src/test/java/de/dhbw/karlsruhe/students/mailflow/integration/LoginServiceTest.java
@@ -4,11 +4,11 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import de.dhbw.karlsruhe.students.mailflow.core.application.auth.LoginService;
 import de.dhbw.karlsruhe.students.mailflow.core.domain.auth.AuthorizationException;
+import de.dhbw.karlsruhe.students.mailflow.core.domain.auth.LoadingUsersException;
 import de.dhbw.karlsruhe.students.mailflow.core.domain.auth.PasswordChecker;
 import de.dhbw.karlsruhe.students.mailflow.core.domain.email.value_objects.Address;
 import de.dhbw.karlsruhe.students.mailflow.core.domain.user.User;
 import de.dhbw.karlsruhe.students.mailflow.core.domain.user.UserRepository;
-import de.dhbw.karlsruhe.students.mailflow.external.infrastructure.authorization.LoadingUsersException;
 import java.util.Optional;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
